LayoutTests/fast/dom/shadow/layout-tests-can-access-shadow-expected.txt


This tests that layout tests can access shadow DOM.

PASS shadow.nodeName is "#shadow-root"
PASS exceptionCode is DOMException.INVALID_ACCESS_ERR
PASS internals.shadowRoot(p) is null
PASS shadow.nodeName is "#shadow-root"
PASS shadow === internals.shadowRoot(p) is true
PASS successfullyParsed is true

TEST COMPLETE
